Prepare token_coop to implement token exchange
There was no way to configure what grant_types to support and token
endpoint had to be refactored to avoid creating another new endpoint
like TokenCoop to support token exchange.

TokenCoop was selected to be refactored but we could in the future
just keep the new version instead of having both Token and the new
configurable/extendable TokenCoop.

Grant types will be configured by the grant_types_supported provider
configuration. Currently this is configured in the authorization
endpoint but maybe we should have a global one.